Re-organize helper widget source code
