+++ lib/flutter_widget_from_html.dart
@@ -2,6 +2,4 @@ export 'src/builder.dart';
 export 'src/data_classes.dart';
 export 'src/helpers.dart';
 export 'src/html_widget.dart';
-export 'src/video_player.dart';
 export 'src/widget_factory.dart';
-export 'src/web_view.dart';
