+++ lib/src/helpers.dart
@@ -1,1 +1,11 @@
 export 'package:flutter_widget_from_html_core/src/core_helpers.dart';
+
+import 'dart:async';
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:chewie/chewie.dart' as lib;
+import 'package:video_player/video_player.dart' as lib;
+import 'package:webview_flutter/webview_flutter.dart' as lib;
+
+part 'widget/video_player.dart';
+part 'widget/web_view.dart';
