+++ lib/src/widget/video_player.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/material.dart';
-import 'package:chewie/chewie.dart' as lib;
-import 'package:video_player/video_player.dart' as lib;
+part of '../helpers.dart';
 
 class VideoPlayer extends StatefulWidget {
   final String url;
